Added barrelling for volume bar
